Removed ToolTip from tkintertable
In previous versions, if you click a cell, a light yellow popup appeared next to the cell. We can edit the cell and the popup is distracting. Occasionally it would capture itself in a cell and crash the program. Gone is that tooltip 
Create a subclass to disable the tooltip
class NoTooltipTableCanvas(TableCanvas):
    def drawTooltip(self, row, col):
        """Disable tooltip"""
        return