Sample file for testing html
try.csv is a csv file for Webster123 with empty fields except for ourUrl, folderPath, and fileExtension. It will publish a blank html page. Html can be put into ourContent, ourHeader, or ourFooter to experiment with html and quickly see the result.
So far Webster123 reads a csv file and makes pages. Individual cells can be edited.
In it's present form, the variable names themselves come in from the csv. Columns can be moved around. In it's final form try.csv is already there. This would make our columns and csv fixed. 
In a final version new records can be entered directly and the csv can be a convenient way to import and export data.
The file try.csv could easily be used to test php as well since we have a cell with a hook to the <html> tag.