Table Editing is Achieving Interactive Use
We've abandoned tkintertables and are editing withing our own table. We can copy and paste selections, insert and delete rows. We are cropping data when presenting it to the screen. This makes Webster123 very fast. It's a fixed table. You double click to edit a cell. Clicking and writing directly in the cell, we don't allow.  At the front of the program we can provide column widths, the maximum number of characters to display in the column, the instructions for the inputs for each input and the list of variables as well as the order to place them. This program can be modified for other uses with other tables.