Version 1.0.0
Entering row numbers was tedious. This version publishes by highlighting rows. Double click to edit a cell in a new window on the taskbar. This enables a window to compare code as composed with ChatGpt or other sources. 
To publish a row (web page), you simply select the row and hit the publish button. The row is already highlighted so the view button views the range in the default browser.
This change makes Webster123 super fast and interactive.
Other changes in this release:
1. Added viewport meta in the <head> of generated pages and utf-8 declaration.
2. Adjusted the width of rows and moved some columns around.
3. Revised sample csv files to correctlu use flex boxes and be mobile.
4. Csv files provide a good way to share templates.
5. Finalized webster1.css, our sample css which is used in webster.csv and can be used by users.
6. Added the menu script to the generated html. 
7. Added "Crafted by Webster123" with a link to webster123.com on every page generated. This can be removed in a pro version for $. Webster123 can now be free.