webster123-v1.0.3.py
Revised table setup at the front of the program. Removed the specification of the number of characters to be shown in each column. That number is calculated based upon font sixe.
Added the ability to specify font, font size and characteristics and the colors of the table. Added colors to the table.
Added a logo, webster-logo-web.png which has to be in the same directory as webster123-v1.0.3.py when using pyinstaller and in the same directory as the exe when packaged by Inno

Signed-off-by: David Michael Jackson <jacksdm@att.net>