Backup is gone and New is in
Removed Backup Database and added New Database to the File Menu. Since we have Commit Database to commit the changes to the array, it's unsure which to backup. Besides we are creating a backup safety net. We backup before we commit. Not a normal db situation but perfect for Webster.

Signed-off-by: David Michael Jackson <jacksdm@att.net>